Combobox & move combobox tests to tests-file
